Split NABERS User Inputs and private figures
