+++ openfisca_nsw_ess_nabers/variables/energy_saving_scheme/NABERS_offices/coefficient_values.py
@@ -2,6 +2,7 @@
 from openfisca_core.model_api import *
 # Import the Entities specifically defined for this tax and benefit system
 from openfisca_nsw_base.entities import *
+from openfisca_nsw_ess_nabers.variables.energy_saving_scheme.NABERS_offices import nabers_office_shared_library_of_terms
 import numpy as np
 
 # measured_electricity_consumption input at Step 1
@@ -41,4 +42,4 @@ CDD_coefficients = {
                     }
 
 def has_real_values():
-    return False
+    return True
